handle errors with toast and 404, update patient features schema, update model, pca, scaler, train embeddings and added notebook

+++ model-api/schema/patient.py
@@ -3,10 +3,10 @@ import pandas as pd
 
 class Patient(BaseModel):
   pregnancies: int
-  glucose: int
-  bloodPressure: int
-  skinThickness: int
-  insulin: int
+  glucose: float
+  bloodPressure: float
+  skinThickness: float
+  insulin: float
   bmi: float
   diabetesPedigreeFunction: float
   age: int
